Revert "修改卡池ID类型为long"
This reverts commit 35537b04406568a6b210e5b7eb34c739c9a4f46a.

+++ Actions/CardReference.cs
@@ -4,7 +4,7 @@ namespace TouhouCardEngine
     [Serializable]
     public class CardReference
     {
-        public long cardPoolId;
+        public int cardPoolId;
         public int cardId;
     }
 }
